test(velvet): share the Motion enter/orchestration fixtures' panel scaffolding
MotionOrchestrationTests and MotionStandaloneEnterTests each hand-rolled the
same ~35 lines of simulated-panel setup (EditorPanelSimulator + Reconciler
creation/disposal, Root/Tick/AdvancePast) needed because their scheduled
swaps only fire once a panel ticks its own clock. Extract
MotionSimulatedPanelTestsBase (mirroring VariantCleanupTestsBase's in-asmdef
sharing, and ReconcilerTestFixture's virtual-SetUp/TearDown convention) and
have both fixtures inherit it instead of duplicating the setup.

Also rewrite MotionSpringDriverTests' panel-free doc comment to state its
own reasoning in self-contained behavioral terms, dropping the out-of-repo
"assignment" reference.
